refactor: StudyUpdateService -> StudyService 통합 및 StudyUpdateRequest 수정
