+++ src/docs/asciidoc/study-update.adoc
@@ -5,88 +5,52 @@
 :toc: left
 :toclevels: 2
 
-== 스터디 정보수정 폼 제공
-=== 1. Authorization Header에 AccessToken이 없으면 스터디 정보수정 폼 제공을 실패한다
-HTTP Request
-include::{snippets}/StudyApi/Update/Get/Failure/Case1/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Get/Failure/Case1/path-parameters.adoc[]
-
-HTTP Response
-include::{snippets}/StudyApi/Update/Get/Failure/Case1/response-body.adoc[]
-
-=== 2. 스터디 팀장이 아니라면 스터디 수정 폼을 제공할 수 없습니다
-HTTP Request
-include::{snippets}/StudyApi/Update/Get/Failure/Case2/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Get/Failure/Case2/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Get/Failure/Case2/path-parameters.adoc[]
-
-HTTP Response
-include::{snippets}/StudyApi/Update/Get/Failure/Case2/response-body.adoc[]
-
-=== 3. 정보 수정 폼 제공에 성공한다 - 온라인
-HTTP Request
-include::{snippets}/StudyApi/Update/Get/Success/Case1/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Get/Success/Case1/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Get/Success/Case1/path-parameters.adoc[]
-
-HTTP Response
-include::{snippets}/StudyApi/Update/Get/Success/Case1/response-body.adoc[]
-
-=== 4. 정보 수정 폼 제공에 성공한다 - 오프라인
-HTTP Request
-include::{snippets}/StudyApi/Update/Get/Success/Case2/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Get/Success/Case2/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Get/Success/Case2/path-parameters.adoc[]
-
-HTTP Response
-include::{snippets}/StudyApi/Update/Get/Success/Case2/response-body.adoc[]
-
 == 스터디 정보수정
 === 1. Authorization Header에 AccessToken이 없으면 스터디 정보수정을 실패한다
 HTTP Request
-include::{snippets}/StudyApi/Update/Put/Failure/Case1/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case1/path-parameters.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case1/request-fields.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case1/http-request.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case1/path-parameters.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case1/request-fields.adoc[]
 
 HTTP Response
-include::{snippets}/StudyApi/Update/Put/Failure/Case1/response-body.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case1/response-body.adoc[]
 
 === 2. 스터디 팀장이 아니라면 정보를 수정할 수 없다
 HTTP Request
-include::{snippets}/StudyApi/Update/Put/Failure/Case2/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case2/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case2/path-parameters.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case2/request-fields.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case2/http-request.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case2/request-headers.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case2/path-parameters.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case2/request-fields.adoc[]
 
 HTTP Response
-include::{snippets}/StudyApi/Update/Put/Failure/Case2/response-body.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case2/response-body.adoc[]
 
 === 3. 최대 수용인원을 현재 스터디 인원보다 적게 설정할 수 없다
 HTTP Request
-include::{snippets}/StudyApi/Update/Put/Failure/Case3/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case3/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case3/path-parameters.adoc[]
-include::{snippets}/StudyApi/Update/Put/Failure/Case3/request-fields.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case3/http-request.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case3/request-headers.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case3/path-parameters.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case3/request-fields.adoc[]
 
 HTTP Response
-include::{snippets}/StudyApi/Update/Put/Failure/Case3/response-body.adoc[]
+include::{snippets}/StudyApi/Update/Failure/Case3/response-body.adoc[]
 
 === 4. 정보 수정에 성공한다 - 온라인
 HTTP Request
-include::{snippets}/StudyApi/Update/Put/Success/Case1/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case1/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case1/path-parameters.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case1/request-fields.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case1/http-request.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case1/request-headers.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case1/path-parameters.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case1/request-fields.adoc[]
 
 HTTP Response
-include::{snippets}/StudyApi/Update/Put/Success/Case1/http-response.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case1/http-response.adoc[]
 
 === 5. 정보 수정에 성공한다 - 오프라인
 HTTP Request
-include::{snippets}/StudyApi/Update/Put/Success/Case2/http-request.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case2/request-headers.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case2/path-parameters.adoc[]
-include::{snippets}/StudyApi/Update/Put/Success/Case2/request-fields.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case2/http-request.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case2/request-headers.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case2/path-parameters.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case2/request-fields.adoc[]
 
 HTTP Response
-include::{snippets}/StudyApi/Update/Put/Success/Case2/http-response.adoc[]
+include::{snippets}/StudyApi/Update/Success/Case2/http-response.adoc[]
