+++ src/main/java/com/kgu/studywithme/study/controller/StudyApiController.java
@@ -2,9 +2,8 @@ package com.kgu.studywithme.study.controller;
 
 import com.kgu.studywithme.global.annotation.ExtractPayload;
 import com.kgu.studywithme.study.controller.dto.request.StudyRegisterRequest;
-import com.kgu.studywithme.study.controller.dto.request.StudyUpdate;
+import com.kgu.studywithme.study.controller.dto.request.StudyUpdateRequest;
 import com.kgu.studywithme.study.service.StudyService;
-import com.kgu.studywithme.study.service.StudyUpdateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,9 +16,8 @@ import javax.validation.Valid;
 @RequestMapping("/api")
 public class StudyApiController {
     private final StudyService studyService;
-    private final StudyUpdateService studyUpdateService;
 
-    @PostMapping
+    @PostMapping("/study")
     public ResponseEntity<Void> register(@RequestBody @Valid StudyRegisterRequest request, @ExtractPayload Long hostId) {
         Long savedStudyId = studyService.register(request, hostId);
         return ResponseEntity
@@ -27,17 +25,11 @@ public class StudyApiController {
                 .build();
     }
 
-    @GetMapping("/{studyId}")
-    public ResponseEntity<StudyUpdate> getUpdateForm(@PathVariable Long studyId, @ExtractPayload Long hostId) {
-        StudyUpdate response = studyUpdateService.getUpdateForm(studyId, hostId);
-        return ResponseEntity.ok(response);
-    }
-
-    @PutMapping("/{studyId}")
+    @PatchMapping("/studies/{studyId}")
     public ResponseEntity<Void> update(@PathVariable Long studyId,
-                                       @RequestBody @Valid StudyUpdate request,
+                                       @RequestBody @Valid StudyUpdateRequest request,
                                        @ExtractPayload Long hostId) {
-        studyUpdateService.update(studyId, request, hostId);
+        studyService.update(studyId, request, hostId);
         return ResponseEntity.noContent().build();
     }
 }
