+++ src/test/java/com/kgu/studywithme/study/controller/StudyApiControllerTest.java
@@ -4,9 +4,8 @@ import com.kgu.studywithme.auth.exception.AuthErrorCode;
 import com.kgu.studywithme.common.ControllerTest;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import com.kgu.studywithme.study.controller.dto.request.StudyRegisterRequest;
-import com.kgu.studywithme.study.controller.dto.request.StudyUpdate;
+import com.kgu.studywithme.study.controller.dto.request.StudyUpdateRequest;
 import com.kgu.studywithme.study.controller.utils.StudyRegisterRequestUtils;
-import com.kgu.studywithme.study.domain.RecruitmentStatus;
 import com.kgu.studywithme.study.exception.StudyErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -36,9 +35,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @DisplayName("Study [Controller Layer] -> StudyApiController 테스트")
 class StudyApiControllerTest extends ControllerTest {
     @Nested
-    @DisplayName("스터디 생성 API [POST /api]")
+    @DisplayName("스터디 생성 API [POST /api/study]")
     class register {
-        private static final String BASE_URL = "/api";
+        private static final String BASE_URL = "/api/study";
 
         @Test
         @DisplayName("Authorization Header에 AccessToken이 없으면 스터디 생성을 실패한다")
@@ -318,212 +317,18 @@ class StudyApiControllerTest extends ControllerTest {
     }
 
     @Nested
-    @DisplayName("수정 폼 제공 API [GET /api/{studyId}]")
-    class showUpdateForm {
-        private static final String BASE_URL = "/api/{studyId}";
-        private static final Long STUDY_ID = 1L;
-
-        @Test
-        @DisplayName("Authorization Header에 AccessToken이 없으면 수정 폼 제공을 실패한다")
-        void withoutAccessToken() throws Exception {
-            // when
-            MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .get(BASE_URL, STUDY_ID);
-
-            // then
-            final AuthErrorCode expectedError = AuthErrorCode.INVALID_PERMISSION;
-            mockMvc.perform(requestBuilder)
-                    .andExpectAll(
-                            status().isForbidden(),
-                            jsonPath("$.status").exists(),
-                            jsonPath("$.status").value(expectedError.getStatus().value()),
-                            jsonPath("$.errorCode").exists(),
-                            jsonPath("$.errorCode").value(expectedError.getErrorCode()),
-                            jsonPath("$.message").exists(),
-                            jsonPath("$.message").value(expectedError.getMessage())
-                    )
-                    .andDo(
-                            document(
-                                    "StudyApi/Update/Get/Failure/Case1",
-                                    getDocumentRequest(),
-                                    getDocumentResponse(),
-                                    pathParameters(
-                                            parameterWithName("studyId").description("수정할 스터디 ID(PK)")
-                                    ),
-                                    responseFields(
-                                            fieldWithPath("status").description("HTTP 상태 코드"),
-                                            fieldWithPath("errorCode").description("커스텀 예외 코드"),
-                                            fieldWithPath("message").description("예외 메시지")
-                                    )
-                            )
-                    );
-        }
-
-        @Test
-        @DisplayName("스터디 팀장이 아니라면 스터디 수정 폼을 제공할 수 없습니다")
-        void throwExceptionByMemberNotHost() throws Exception {
-            // given
-            given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
-            given(jwtTokenProvider.getId(anyString())).willReturn(1L);
-            doThrow(StudyWithMeException.type(StudyErrorCode.MEMBER_IS_NOT_HOST))
-                    .when(studyUpdateService)
-                    .getUpdateForm(any(), any());
-
-            // when
-            MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .get(BASE_URL, STUDY_ID)
-                    .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN));
-
-            // then
-            final StudyErrorCode expectedError = StudyErrorCode.MEMBER_IS_NOT_HOST;
-            mockMvc.perform(requestBuilder)
-                    .andExpectAll(
-                            status().isConflict(),
-                            jsonPath("$.status").exists(),
-                            jsonPath("$.status").value(expectedError.getStatus().value()),
-                            jsonPath("$.errorCode").exists(),
-                            jsonPath("$.errorCode").value(expectedError.getErrorCode()),
-                            jsonPath("$.message").exists(),
-                            jsonPath("$.message").value(expectedError.getMessage())
-                    )
-                    .andDo(
-                            document(
-                                    "StudyApi/Update/Get/Failure/Case2",
-                                    getDocumentRequest(),
-                                    getDocumentResponse(),
-                                    requestHeaders(
-                                            headerWithName(AUTHORIZATION).description("Access Token")
-                                    ),
-                                    pathParameters(
-                                            parameterWithName("studyId").description("수정할 스터디 ID(PK)")
-                                    ),
-                                    responseFields(
-                                            fieldWithPath("status").description("HTTP 상태 코드"),
-                                            fieldWithPath("errorCode").description("커스텀 예외 코드"),
-                                            fieldWithPath("message").description("예외 메시지")
-                                    )
-                            )
-                    );
-        }
-
-        @Test
-        @DisplayName("정보 수정 폼 제공에 성공한다 - 온라인")
-        void successOnline() throws Exception {
-            // given
-            final StudyUpdate response = generateOnlineStudyUpdate();
-            given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
-            given(jwtTokenProvider.getId(anyString())).willReturn(1L);
-            doReturn(response)
-                    .when(studyUpdateService)
-                    .getUpdateForm(any(), any());
-
-            // when
-            MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .get(BASE_URL, STUDY_ID)
-                    .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN));
-
-            // then
-            mockMvc.perform(requestBuilder)
-                    .andExpectAll(
-                            status().isOk()
-                    )
-                    .andDo(
-                            document(
-                                    "StudyApi/Update/Get/Success/Case1",
-                                    getDocumentRequest(),
-                                    getDocumentResponse(),
-                                    requestHeaders(
-                                            headerWithName(AUTHORIZATION).description("Access Token")
-                                    ),
-                                    pathParameters(
-                                            parameterWithName("studyId").description("수정할 스터디 ID(PK)")
-                                    ),
-                                    responseFields(
-                                            fieldWithPath("name").description("스터디명"),
-                                            fieldWithPath("description").description("스터디 설명"),
-                                            fieldWithPath("category").description("스터디 카테고리"),
-                                            fieldWithPath("capacity").description("스터디 최대 인원"),
-                                            fieldWithPath("type").description("스터디 타입")
-                                                    .attributes(constraint("온라인 / 오프라인")),
-                                            fieldWithPath("province").description("오프라인 스터디 지역 [경기도, 강원도, ...]")
-                                                    .optional()
-                                                    .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
-                                                    .optional()
-                                                    .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
-                                            fieldWithPath("hashtags[]").description("스터디 해시태그")
-                                    )
-                            )
-                    );
-        }
-
-        @Test
-        @DisplayName("정보 수정 폼 제공에 성공한다 - 오프라인")
-        void successOffline() throws Exception {
-            // given
-            final StudyUpdate response = generateOfflineStudyUpdate();
-            given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
-            given(jwtTokenProvider.getId(anyString())).willReturn(1L);
-            doReturn(response)
-                    .when(studyUpdateService)
-                    .getUpdateForm(any(), any());
-
-            // when
-            MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .get(BASE_URL, STUDY_ID)
-                    .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN));
-
-            // then
-            mockMvc.perform(requestBuilder)
-                    .andExpectAll(
-                            status().isOk()
-                    )
-                    .andDo(
-                            document(
-                                    "StudyApi/Update/Get/Success/Case2",
-                                    getDocumentRequest(),
-                                    getDocumentResponse(),
-                                    requestHeaders(
-                                            headerWithName(AUTHORIZATION).description("Access Token")
-                                    ),
-                                    pathParameters(
-                                            parameterWithName("studyId").description("수정할 스터디 ID(PK)")
-                                    ),
-                                    responseFields(
-                                            fieldWithPath("name").description("스터디명"),
-                                            fieldWithPath("description").description("스터디 설명"),
-                                            fieldWithPath("category").description("스터디 카테고리"),
-                                            fieldWithPath("capacity").description("스터디 최대 인원"),
-                                            fieldWithPath("type").description("스터디 타입")
-                                                    .attributes(constraint("온라인 / 오프라인")),
-                                            fieldWithPath("province").description("오프라인 스터디 지역 [경기도, 강원도, ...]")
-                                                    .optional()
-                                                    .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
-                                                    .optional()
-                                                    .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
-                                            fieldWithPath("hashtags[]").description("스터디 해시태그")
-                                    )
-                            )
-                    );
-        }
-    }
-
-    @Nested
-    @DisplayName("정보 수정 API [PUT /api/{studyId}]")
+    @DisplayName("정보 수정 API [PATCH /api/studies/{studyId}]")
     class update {
-        private static final String BASE_URL = "/api/{studyId}";
+        private static final String BASE_URL = "/api/studies/{studyId}";
         private static final String STUDY_ID = "1";
 
         @Test
         @DisplayName("Authorization Header에 AccessToken이 없으면 정보 수정을 실패한다")
         void withoutAccessToken() throws Exception {
             // when
-            final StudyUpdate request = generateOfflineStudyUpdate();
+            final StudyUpdateRequest request = generateOfflineStudyUpdate();
             MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .put(BASE_URL, STUDY_ID)
+                    .patch(BASE_URL, STUDY_ID)
                     .contentType(APPLICATION_JSON)
                     .content(convertObjectToJson(request));
 
@@ -541,7 +346,7 @@ class StudyApiControllerTest extends ControllerTest {
                     )
                     .andDo(
                             document(
-                                    "StudyApi/Update/Put/Failure/Case1",
+                                    "StudyApi/Update/Failure/Case1",
                                     getDocumentRequest(),
                                     getDocumentResponse(),
                                     pathParameters(
@@ -559,7 +364,7 @@ class StudyApiControllerTest extends ControllerTest {
                                             fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
                                                     .optional()
                                                     .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
+                                            fieldWithPath("status").description("스터디 모집 여부"),
                                             fieldWithPath("hashtags").description("해시태그")
                                     ),
                                     responseFields(
@@ -578,13 +383,13 @@ class StudyApiControllerTest extends ControllerTest {
             given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
             given(jwtTokenProvider.getId(anyString())).willReturn(1L);
             doThrow(StudyWithMeException.type(StudyErrorCode.MEMBER_IS_NOT_HOST))
-                    .when(studyUpdateService)
+                    .when(studyService)
                     .update(any(), any(), any());
 
             // when
-            final StudyUpdate request = generateOfflineStudyUpdate();
+            final StudyUpdateRequest request = generateOfflineStudyUpdate();
             MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .put(BASE_URL, STUDY_ID)
+                    .patch(BASE_URL, STUDY_ID)
                     .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN))
                     .contentType(APPLICATION_JSON)
                     .content(convertObjectToJson(request));
@@ -603,7 +408,7 @@ class StudyApiControllerTest extends ControllerTest {
                     )
                     .andDo(
                             document(
-                                    "StudyApi/Update/Put/Failure/Case2",
+                                    "StudyApi/Update/Failure/Case2",
                                     getDocumentRequest(),
                                     getDocumentResponse(),
                                     requestHeaders(
@@ -624,7 +429,7 @@ class StudyApiControllerTest extends ControllerTest {
                                             fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
                                                     .optional()
                                                     .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
+                                            fieldWithPath("status").description("스터디 모집 여부"),
                                             fieldWithPath("hashtags").description("해시태그")
                                     ),
                                     responseFields(
@@ -643,13 +448,13 @@ class StudyApiControllerTest extends ControllerTest {
             given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
             given(jwtTokenProvider.getId(anyString())).willReturn(1L);
             doThrow(StudyWithMeException.type(StudyErrorCode.CAPACITY_CANNOT_BE_LESS_THAN_MEMBERS))
-                    .when(studyUpdateService)
+                    .when(studyService)
                     .update(any(), any(), any());
 
             // when
-            final StudyUpdate request = generateOfflineStudyUpdate();
+            final StudyUpdateRequest request = generateOfflineStudyUpdate();
             MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .put(BASE_URL, STUDY_ID)
+                    .patch(BASE_URL, STUDY_ID)
                     .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN))
                     .contentType(APPLICATION_JSON)
                     .content(convertObjectToJson(request));
@@ -668,7 +473,7 @@ class StudyApiControllerTest extends ControllerTest {
                     )
                     .andDo(
                             document(
-                                    "StudyApi/Update/Put/Failure/Case3",
+                                    "StudyApi/Update/Failure/Case3",
                                     getDocumentRequest(),
                                     getDocumentResponse(),
                                     requestHeaders(
@@ -689,7 +494,7 @@ class StudyApiControllerTest extends ControllerTest {
                                             fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
                                                     .optional()
                                                     .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
+                                            fieldWithPath("status").description("스터디 모집 여부"),
                                             fieldWithPath("hashtags").description("해시태그")
                                     ),
                                     responseFields(
@@ -708,13 +513,13 @@ class StudyApiControllerTest extends ControllerTest {
             given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
             given(jwtTokenProvider.getId(anyString())).willReturn(1L);
             doNothing()
-                    .when(studyUpdateService)
+                    .when(studyService)
                     .update(any(), any(), any());
 
             // when
-            final StudyUpdate request = generateOnlineStudyUpdate();
+            final StudyUpdateRequest request = generateOnlineStudyUpdate();
             MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .put(BASE_URL, STUDY_ID)
+                    .patch(BASE_URL, STUDY_ID)
                     .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN))
                     .contentType(APPLICATION_JSON)
                     .content(convertObjectToJson(request));
@@ -726,7 +531,7 @@ class StudyApiControllerTest extends ControllerTest {
                     )
                     .andDo(
                             document(
-                                    "StudyApi/Update/Put/Success/Case1",
+                                    "StudyApi/Update/Success/Case1",
                                     getDocumentRequest(),
                                     getDocumentResponse(),
                                     requestHeaders(
@@ -747,7 +552,7 @@ class StudyApiControllerTest extends ControllerTest {
                                             fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
                                                     .optional()
                                                     .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
+                                            fieldWithPath("status").description("스터디 모집 여부"),
                                             fieldWithPath("hashtags").description("해시태그")
                                     )
                             )
@@ -761,13 +566,13 @@ class StudyApiControllerTest extends ControllerTest {
             given(jwtTokenProvider.isTokenValid(anyString())).willReturn(true);
             given(jwtTokenProvider.getId(anyString())).willReturn(1L);
             doNothing()
-                    .when(studyUpdateService)
+                    .when(studyService)
                     .update(any(), any(), any());
 
             // when
-            final StudyUpdate request = generateOfflineStudyUpdate();
+            final StudyUpdateRequest request = generateOfflineStudyUpdate();
             MockHttpServletRequestBuilder requestBuilder = RestDocumentationRequestBuilders
-                    .put(BASE_URL, STUDY_ID)
+                    .patch(BASE_URL, STUDY_ID)
                     .header(AUTHORIZATION, String.join(" ", BEARER_TOKEN, ACCESS_TOKEN))
                     .contentType(APPLICATION_JSON)
                     .content(convertObjectToJson(request));
@@ -779,7 +584,7 @@ class StudyApiControllerTest extends ControllerTest {
                     )
                     .andDo(
                             document(
-                                    "StudyApi/Update/Put/Success/Case2",
+                                    "StudyApi/Update/Success/Case2",
                                     getDocumentRequest(),
                                     getDocumentResponse(),
                                     requestHeaders(
@@ -800,7 +605,7 @@ class StudyApiControllerTest extends ControllerTest {
                                             fieldWithPath("city").description("오프라인 스터디 지역 [안양시, 수원시, ...]")
                                                     .optional()
                                                     .attributes(constraint("오프라인 스터디의 경우 필수")),
-                                            fieldWithPath("recruitmentStatus").description("스터디 모집 여부"),
+                                            fieldWithPath("status").description("스터디 모집 여부"),
                                             fieldWithPath("hashtags").description("해시태그")
                                     )
                             )
@@ -808,8 +613,8 @@ class StudyApiControllerTest extends ControllerTest {
         }
     }
 
-    public StudyUpdate generateOfflineStudyUpdate() {
-        return StudyUpdate.builder()
+    public StudyUpdateRequest generateOfflineStudyUpdate() {
+        return StudyUpdateRequest.builder()
                 .name(TOSS_INTERVIEW.getName())
                 .description(TOSS_INTERVIEW.getDescription())
                 .capacity(TOSS_INTERVIEW.getCapacity())
@@ -817,13 +622,13 @@ class StudyApiControllerTest extends ControllerTest {
                 .type(TOSS_INTERVIEW.getType().getDescription())
                 .province(TOSS_INTERVIEW.getArea().getProvince())
                 .city(TOSS_INTERVIEW.getArea().getCity())
-                .recruitmentStatus(RecruitmentStatus.COMPLETE.getDescription())
+                .status(true)
                 .hashtags(TOSS_INTERVIEW.getHashtags())
                 .build();
     }
 
-    public StudyUpdate generateOnlineStudyUpdate() {
-        return StudyUpdate.builder()
+    public StudyUpdateRequest generateOnlineStudyUpdate() {
+        return StudyUpdateRequest.builder()
                 .name(TOEFL.name())
                 .description(TOEFL.getDescription())
                 .capacity(TOEFL.getCapacity())
@@ -831,7 +636,7 @@ class StudyApiControllerTest extends ControllerTest {
                 .type(TOEFL.getType().getDescription())
                 .province(null)
                 .city(null)
-                .recruitmentStatus(RecruitmentStatus.COMPLETE.getDescription())
+                .status(true)
                 .hashtags(TOEFL.getHashtags())
                 .build();
     }
